napoveda

+++ game.js
@@ -705,36 +705,6 @@ function handlePuzzleInteraction(item) {
 	}
 }
 
-/**
- * Zpracuje řešení hádanky.
- */
-/* window.solvePuzzle = function () {
-	const sym1 = document.getElementById('symbol1').value.toUpperCase();
-	const sym2 = document.getElementById('symbol2').value.toUpperCase();
-	const sym3 = document.getElementById('symbol3').value.toUpperCase();
-	const sym4 = document.getElementById('symbol4').value.toUpperCase();
-
-	// Správné řešení: KLIC, LIST, KRIZ, PEČET
-	if (sym1 === 'ERB' && sym2 === 'LIST' && (sym3 === 'KRIZ' || sym3 == 'KŘÍŽ') && sym4 === 'KORUNA') {
-
-		// Otevření nové cesty do Jeskyně
-		MAP['kamenna_tvar']['E'].forward = 'jeskyne';
-
-		hidePopup('puzzle-modal-backdrop');
-		showPopup('Mechanismus zaskřípe a s hlasitým duněním se balvan odsune. <br><br><span class="color-lighter-red">Objevil se tajný vstup do jeskyně!</span><br><br> Nyní se musím pouze odhodlat a jít vpřed.');
-
-		// Zrušení hádanky, už není potřeba
-		const puzzleHotspot = MAP['kamenna_tvar']['N'].items.find(i => i.type === 'puzzle');
-		if (puzzleHotspot) puzzleHotspot.text = "Vstup do jeskyně je nyní otevřen.";
-
-		updateView();
-
-	} else {
-		showPopup('Nic se nestalo.');
-	}
-} */
-
-
 
 /**
 * Správný symbol je určen rotací kola a je vyčten z jeho datového atributu.
@@ -997,6 +967,7 @@ window.onload = function () {
 	let ovladaci_prvky = document.getElementById('ovladaci_prvky');
 
 	let start_btn = document.getElementById('spustit_hru');
+	let load_btn_start = document.getElementById('nacist_hru_start');
 	let load_btn = document.getElementById('nacist_hru');
 	let save_btn = document.getElementById('ulozit_hru');
 
@@ -1021,48 +992,38 @@ window.onload = function () {
 	};
 
 	/* LOAD GAME*/
+
+	load_btn_start.onclick = () => {
+		nacistHru();
+	};
+
 	load_btn.onclick = () => {
-		
-		let saved_game_json = this.localStorage.getItem("saved_game");
-
-		if(saved_game_json.length == 0) {
-			alert("Žádná uložená hra k načtení.");
-			return;
-		}else{
-
-			startScreen.style.display = 'none';
-
-			let saved_game = JSON.parse(saved_game_json);
-
-			currentArea = saved_game.currentArea;
-			currentDirectionIndex = saved_game.currentDirectionIndex;
-			inventory = saved_game.inventory;
-			solvedPuzzles = saved_game.solvedPuzzles;
-			addedItems = saved_game.addedItems;
-
-			this.localStorage.setItem("hra_spustena", 1);
-
-			ovladaci_prvky.style.display = 'block';
-
-			updateView();
+		let potvrzeni_nacteni = this.confirm("Chcete načíst uloženou hru?");
+		if(potvrzeni_nacteni) {
+			nacistHru();
 		}
-
-		
 	};
 
 	/* SAVE GAME */
 	save_btn.onclick = () => {
-		let saved_game = {
-			currentArea: currentArea,
-			currentDirectionIndex: currentDirectionIndex,
-			inventory: inventory,
-			solvedPuzzles: solvedPuzzles,
-			addedItems: addedItems
-		};
-		this.localStorage.setItem("saved_game", JSON.stringify(saved_game));
 
-		if(this.localStorage.getItem("saved_game").length > 0) {
-			alert("Hra byla uložena.");
+		let potvrzeni_ulozeni = this.confirm("Chcete uložit hru?");
+
+		if(potvrzeni_ulozeni) {
+
+			let saved_game = {
+				currentArea: currentArea,
+				currentDirectionIndex: currentDirectionIndex,
+				inventory: inventory,
+				solvedPuzzles: solvedPuzzles,
+				addedItems: addedItems
+			};
+
+			this.localStorage.setItem("saved_game", JSON.stringify(saved_game));
+
+			if(this.localStorage.getItem("saved_game").length > 0) {
+				alert("Hra byla uložena.");
+			}
 		}
 	};
 
@@ -1090,7 +1051,49 @@ window.onload = function () {
 		denik.style.display = 'none';
 	};
 
+	/* NAPOVEDA */
+
+	let napoveda_link = document.getElementById('napoveda_link');
+	napoveda_link.onclick = () => {
+		let napoveda = document.getElementById('napoveda');
+		napoveda.style.display = 'flex';
+	};
+
+	/* zavrit napoveda */
+	let close_napoveda = document.getElementById('zavrit_napoveda');
+	close_napoveda.onclick = () => {
+		let napoveda = document.getElementById('napoveda');
+		napoveda.style.display = 'none';
+	};
+
 };
+
+function nacistHru(){
+
+	let saved_game_json = this.localStorage.getItem("saved_game");
+
+	if(saved_game_json.length == 0) {
+		alert("Žádná uložená hra k načtení.");
+		return;
+	}else{
+
+		startScreen.style.display = 'none';
+
+		let saved_game = JSON.parse(saved_game_json);
+
+		currentArea = saved_game.currentArea;
+		currentDirectionIndex = saved_game.currentDirectionIndex;
+		inventory = saved_game.inventory;
+		solvedPuzzles = saved_game.solvedPuzzles;
+		addedItems = saved_game.addedItems;
+
+		this.localStorage.setItem("hra_spustena", 1);
+
+		ovladaci_prvky.style.display = 'block';
+
+		updateView();
+	}
+}
 
 window.addEventListener('beforeunload', function (e) {
